requirements.txt
# Market Inventory Management System Dependencies
# Python packages required for the application

# Database connectivity
pymysql==1.1.1

# Environment variable management
python-dotenv==1.1.1

# Table formatting for reports
tabulate==0.9.0

# Additional dependencies for enhanced functionality
# (These are optional but recommended)
# mysql-connector-python==9.4.0  # Alternative MySQL connector 